Conflicts:
	virtual-file-system/src/main/java/com/sos/VirtualFileSystem/FTP/SOSVfsFtp.java

+++ virtual-file-system/src/main/java/com/sos/VirtualFileSystem/DataElements/SOSFileList.java
@@ -145,8 +145,11 @@ public class SOSFileList extends SOSVfsMessageCodes {
 					case transfer_aborted:
 						lngFailedTransfers++;
 						break;
-					case notOverwritten:
+					case setBack:
 						lngFailedTransfers++;
+						break;
+					case notOverwritten:
+						lngSkippedTransfers++;
 						break;
 					case waiting4transfer:
 						lngFailedTransfers++;
@@ -455,11 +458,15 @@ public class SOSFileList extends SOSVfsMessageCodes {
 			if (objOptions.isAtomicTransfer() && objOptions.transactional.isTrue()) {
 				logger.debug(SOSVfs_D_209.get());
 				for (SOSFileListEntry objListItem : objFileListEntries) {
+					if (objListItem.isNotOverwritten()) {
+						continue;
+					}
 					String strTargetTransferName = objListItem.TargetTransferName();
 					String strToFilename = MakeFullPathName(objOptions.TargetDir.Value(), objListItem.TargetFileName());
 					if (!fileNamesAreEqual(strToFilename, strTargetTransferName, false)) { // SOSFTP-142
 						ISOSVirtualFile objF = null;
 						if (objOptions.overwrite_files.isTrue() && objListItem.FileExists() == true) {
+							objListItem.setTargetFileAlreadyExists(true);
 							objF = objDataTargetClient.getFileHandle(strToFilename);
 							objF.delete();
 						}
@@ -570,14 +577,36 @@ public class SOSFileList extends SOSVfsMessageCodes {
 		// TODO löschen der Dateien mit Atomic-Prefix und -Suffix auf dem Target
 		if (objOptions.isAtomicTransfer()) {
 			for (SOSFileListEntry objListItem : objFileListEntries) {
-				String strAtomicFileName = objListItem.getAtomicFileName();
-				if (isNotEmpty(strAtomicFileName)) {
-					objDataTargetClient.getFileHandle(strAtomicFileName).delete();
-					String strT = SOSVfs_D_212.params(strAtomicFileName);
-					logger.debug(strT);
-					objJadeReportLogger.info(strT);
-					objListItem.setAtomicFileName(EMPTY_STRING);
-					objListItem.setStatus(enuTransferStatus.setBack);
+				String strAtomicFileName = MakeFullPathName(objOptions.TargetDir.Value(), objListItem.getAtomicFileName());
+				if (isNotEmpty(objListItem.getAtomicFileName())) {
+					try {
+						ISOSVirtualFile atomicFile = objDataTargetClient.getFileHandle(strAtomicFileName);
+						if(atomicFile.FileExists()) {
+							atomicFile.delete();
+						}
+						String strT = SOSVfs_D_212.params(strAtomicFileName);
+						logger.debug(strT);
+						objJadeReportLogger.info(strT);
+						objListItem.setAtomicFileName(EMPTY_STRING);
+						objListItem.setStatus(enuTransferStatus.setBack);
+					} catch (Exception e) {
+						logger.error(e.getLocalizedMessage());
+					}
+				}
+				if (!objListItem.isTargetFileAlreadyExists()) {
+					try {
+						String strTargetFilename = MakeFullPathName(objOptions.TargetDir.Value(), objListItem.TargetFileName());
+						ISOSVirtualFile targetFile = objDataTargetClient.getFileHandle(strTargetFilename);
+						if(targetFile.FileExists()) {
+							targetFile.delete();
+						}
+						String strT = SOSVfs_D_212.params(targetFile);
+						logger.debug(strT);
+						objJadeReportLogger.info(strT);
+						objListItem.setStatus(enuTransferStatus.setBack);
+					} catch (Exception e) {
+						logger.error(e.getLocalizedMessage());
+					}
 				}
 			}
 		}
